requirements.txt
aiofiles==0.5.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
h11==0.11.0
idna==2.10
Pillow==8.0.1
pydantic==1.6.1
requests==2.24.0
soupsieve==2.0.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.2
